Gebruik long in plaats van int

+++ examenstatus-core/src/main/java/nl/topicus/examenwarroom/data/Instelling.java
@@ -1,9 +1,10 @@
 package nl.topicus.examenwarroom.data;
 
-public class Instelling {
-	public Instelling(String naam, long instellingId, String brin,
-			String adres, String postcode, String plaats,
-			String telefoonnummer, int aantalExamenLeerlingen) {
+public class Instelling
+{
+	public Instelling(String naam, long instellingId, String brin, String adres, String postcode,
+			String plaats, String telefoonnummer, long aantalExamenLeerlingen)
+	{
 		super();
 		this.naam = naam;
 		this.instellingId = instellingId;
@@ -29,69 +30,85 @@ public class Instelling {
 
 	private String telefoonnummer;
 
-	private int aantalExamenLeerlingen;
+	private long aantalExamenLeerlingen;
 
-	public String getNaam() {
+	public String getNaam()
+	{
 		return naam;
 	}
 
-	public void setNaam(String naam) {
+	public void setNaam(String naam)
+	{
 		this.naam = naam;
 	}
 
-	public long getInstellingId() {
+	public long getInstellingId()
+	{
 		return instellingId;
 	}
 
-	public void setInstellingId(long instellingId) {
+	public void setInstellingId(long instellingId)
+	{
 		this.instellingId = instellingId;
 	}
 
-	public String getBrin() {
+	public String getBrin()
+	{
 		return brin;
 	}
 
-	public void setBrin(String brin) {
+	public void setBrin(String brin)
+	{
 		this.brin = brin;
 	}
 
-	public String getAdres() {
+	public String getAdres()
+	{
 		return adres;
 	}
 
-	public void setAdres(String adres) {
+	public void setAdres(String adres)
+	{
 		this.adres = adres;
 	}
 
-	public String getPostcode() {
+	public String getPostcode()
+	{
 		return postcode;
 	}
 
-	public void setPostcode(String postcode) {
+	public void setPostcode(String postcode)
+	{
 		this.postcode = postcode;
 	}
 
-	public String getPlaats() {
+	public String getPlaats()
+	{
 		return plaats;
 	}
 
-	public void setPlaats(String plaats) {
+	public void setPlaats(String plaats)
+	{
 		this.plaats = plaats;
 	}
 
-	public String getTelefoonnummer() {
+	public String getTelefoonnummer()
+	{
 		return telefoonnummer;
 	}
 
-	public void setTelefoonnummer(String telefoonnummer) {
+	public void setTelefoonnummer(String telefoonnummer)
+	{
 		this.telefoonnummer = telefoonnummer;
 	}
 
-	public int getAantalExamenLeerlingen() {
+	public long getAantalExamenLeerlingen()
+	{
 		return aantalExamenLeerlingen;
 	}
 
-	public void setAantalExamenLeerlingen(int aantalExamenLeerlingen) {
+	public void setAantalExamenLeerlingen(long aantalExamenLeerlingen)
+	{
 		this.aantalExamenLeerlingen = aantalExamenLeerlingen;
 	}
 }
